backend/requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn>=0.27.0,<1.0.0
httpx>=0.27.0,<1.0.0
pymupdf>=1.24.0,<2.0.0
python-docx>=1.1.0,<2.0.0
python-multipart>=0.0.9,<1.0.0
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.23.0,<1.0.0
